Extract a Board.js
